+++ cpanel/client/app/libraries/create_new_alertify.js
@@ -1,9 +1,9 @@
 /**
  * Create new alertify
  */
-createNewAlertify = function (names) {
-
+createNewAlertify = function (names, transition) {
     var alerts = _.isArray(names) ? names : [names];
+    transition = _.isUndefined(transition) ? 'fade' : transition;
 
     _.each(alerts, function (element) {
 
@@ -12,18 +12,35 @@ createNewAlertify = function (names) {
         if (!alertify[name]) {
             alertify.dialog(name, function () {
                 return {
+                    main: function (title, message) {
+                        this.setting('title', title);
+                        this.message = message;
+
+                        if (message.html && message.instance) {
+                            this.message = message.html;
+                            this.instance = message.instance;
+                        }
+                    },
                     setup: function () {
                         return {
                             options: {
                                 maximizable: true,
                                 closableByDimmer: false,
                                 resizable: false,
-                                transition: "fade"
+                                transition: transition
                             }
                         };
                     },
                     prepare: function () {
+                        this.setContent(this.message);
                         this.elements.footer.style.visibility = "hidden";
+                    },
+                    hooks: {
+                        onclose: function () {
+                            if (this.instance) {
+                                Blaze.remove(this.instance);
+                            }
+                        }
                     }
                 };
             }, false, 'alert');