+++ sample/client/routers/location.js
@@ -3,10 +3,7 @@ var subs = new SubsManager();
 sampleRoutes.route('/location', {
     name: 'sample.location',
     subscriptions: function (params, queryParams) {
-        this.register(
-            'sample_location',
-            subs.subscribe('sample_location')
-        );
+        //this.register('sample_location', subs.subscribe('sample_location'));
     },
     action: function (params, queryParams) {
         Layout.main('sample_location');