+++ sample/client/templates/customer/customer.js
@@ -5,7 +5,12 @@ var indexTpl = Template.sample_customer,
     insertTpl = Template.sample_customerInsert,
     updateTpl = Template.sample_customerUpdate,
     showTpl = Template.sample_customerShow,
-    locationAddonTpl = Template.sample_locationInsert;
+    locationAddOnTpl = Template.sample_locationAddOnCustomer;
+
+/**
+ * State
+ */
+var state = new ReactiveObj();
 
 /**
  * Index
@@ -18,7 +23,8 @@ indexTpl.onCreated(function () {
     });
 
     // Create new  alertify
-    createNewAlertify(["customer", "location"]);
+    createNewAlertify(["customer"]);
+    createNewAlertify(["locationAddon"], 'zoom');
 });
 
 indexTpl.onRendered(function () {
@@ -33,10 +39,13 @@ indexTpl.events({
             .maximize();
     },
     'click .update': function (e, t) {
-        var data = Sample.Collection.Customer.findOne(this._id);
-
-        alertify.customer(fa("pencil", "Customer"), renderTemplate(updateTpl, data))
-            .maximize();
+        Meteor.call('sample_customerById', this._id, function (error, result) {
+            if (!error) {
+                result.dob = moment(result.dob).format('YYYY-MM-DD');
+                alertify.customer(fa("pencil", "Customer"), renderTemplate(updateTpl, result))
+                    .maximize();
+            }
+        })
     },
     'click .remove': function (e, t) {
         var self = this;
@@ -45,7 +54,7 @@ indexTpl.events({
             fa("remove", "Customer"),
             "Are you sure to delete [" + self._id + "]?",
             function () {
-                Sample.Collection.Customer.softRemove(self._id, function (error) {
+                Sample.Collection.Customer.remove(self._id, function (error) {
                     if (error) {
                         alertify.error(error.message);
                     } else {
@@ -58,15 +67,11 @@ indexTpl.events({
 
     },
     'click .show': function (e, t) {
-        var data = Sample.Collection.Customer.findOne({_id: this._id});
-        data.photoUrl = null;
-
-        if (!_.isUndefined(data.photo)) {
-            data.photoUrl = Files.findOne(data.photo)
-                .url();
-        }
-
-        alertify.alert(fa("eye", "Customer"), renderTemplate(showTpl, data));
+        Meteor.call('sample_customerById', this._id, function (error, result) {
+            if (!error) {
+                alertify.alert(fa("eye", "Customer"), renderTemplate(showTpl, result));
+            }
+        })
     },
     'dblclick tbody > tr': function (event) {
         var dataTable = $(event.target)
@@ -79,37 +84,68 @@ indexTpl.events({
     }
 });
 
-indexTpl.onDestroyed(function () {
-    //
-});
-
 /**
  * Insert
  */
 insertTpl.onRendered(function () {
+    state.set('location', {});
     configOnRender();
 });
 
-insertTpl.helpers({});
+insertTpl.helpers({
+    location: function () {
+        return state.get('location');
+    }
+});
 
 insertTpl.events({
     'click .locationAddon': function (e, t) {
-        alertify.location(fa("plus", "Location"), renderTemplate(locationAddonTpl))
+        alertify.locationAddon(fa("plus", "Location"), renderTemplate(locationAddOnTpl))
+            .maximize();
     }
+});
+
+insertTpl.onDestroyed(function () {
 });
 
 /**
  * Update
  */
 updateTpl.onRendered(function () {
+    var dataUpdate = Template.currentData();
+    state.set('location', {
+        _id: dataUpdate.locationId,
+        name: dataUpdate._location.name
+    });
+
     configOnRender();
 });
 
-updateTpl.helpers({});
+updateTpl.helpers({
+    location: function () {
+        return state.get('location');
+    }
+});
 
 updateTpl.events({
     'click .locationAddon': function (e, t) {
-        alertify.location(fa("plus", "Location"), renderTemplate(locationAddonTpl));
+        alertify.locationAddon(fa("plus", "Location"), renderTemplate(locationAddOnTpl))
+            .maximize();
+    }
+});
+
+/**
+ * Location add on
+ */
+locationAddOnTpl.events({
+    'dblclick tbody > tr': function (event) {
+        var dataTable = $(event.target).closest('table').DataTable();
+        var rowData = dataTable.row(event.currentTarget).data();
+
+        //$('label [for="locationId"]').val('Lcation: ' + rowData._id);
+        //$('[name="locationId"]').val(rowData._id);
+        state.set('location', rowData);
+        alertify.locationAddon().close();
     }
 });
 
@@ -122,7 +158,7 @@ AutoForm.hooks({
         before: {
             insert: function (doc) {
                 var prefix = Session.get('currentBranch') + '-';
-                doc._id = idGenerator.genWithPrefix(Sample.Collection.Customer, prefix, 4);
+                doc._id = idGenerator.genWithPrefix(Sample.Collection.Customer, prefix, 6);
                 doc.cpanel_branchId = Session.get('currentBranch');
                 return doc;
             }
@@ -168,41 +204,42 @@ var configOnRender = function () {
     var dob = $('[name="dob"]');
     DateTimePicker.date(dob);
 
-    $('[name="locationId"]')
-        .select2({
-            placeholder: "Search location",
-            allowClear: true,
-            minimumInputLength: 3,
-            ajax: {
-                data: function (params) {
-                    return params;
-                },
-                transport: function (args) {
-                    // Meteor method call
-                    Meteor.call('school_listAddress', args.data, function (err, results) {
-                        if (err) {
-                            args.error(err);
-                            return;
-                        }
+    // Remote select2 (Meteor method)
+    //$('[name="locationId"]')
+    //    .select2({
+    //        placeholder: "Search location",
+    //        allowClear: true,
+    //        minimumInputLength: 3,
+    //        ajax: {
+    //            data: function (params) {
+    //                return params;
+    //            },
+    //            transport: function (args) {
+    //                // Meteor method call
+    //                Meteor.call('school_listAddress', args.data, function (err, results) {
+    //                    if (err) {
+    //                        args.error(err);
+    //                        return;
+    //                    }
+    //
+    //                    args.success(results);
+    //                });
+    //            },
+    //            results: function (data) {
+    //                var results = [];
+    //                _.each(data, function (result) {
+    //                    results.push({
+    //                        id: result.value,
+    //                        text: result.label
+    //                    });
+    //                });
+    //
+    //                return {results: results};
+    //            }
+    //        }
+    //    });
 
-                        args.success(results);
-                    });
-                },
-                results: function (data) {
-                    var results = [];
-                    _.each(data, function (result) {
-                        results.push({
-                            id: result.value,
-                            text: result.label
-                        });
-                    });
-
-                    return {results: results};
-                }
-            }
-        });
-
-    // Remote select2
+    // Remote select2 (URL)
     //$('[name="locationId"]').select2({
     //    placeholder: "Search location",
     //    allowClear: true,