+++ sample/common/collections/customer.js
@@ -5,9 +5,8 @@ Sample.Collection.Customer = new Mongo.Collection("sample_customer");
 Sample.Schema.Customer = new SimpleSchema({
     name: {
         type: String,
-        label: "Name"
-        //unique: true,
-        //max: 200
+        label: "Name",
+        max: 200
     },
     gender: {
         type: String,
@@ -17,57 +16,14 @@ Sample.Schema.Customer = new SimpleSchema({
             options: function () {
                 return Sample.List.gender();
             }
-            //type: "selectize",
-            //afFieldInput: {
-            //    multiple: true,
-            //    selectizeOptions: {}
-            //}
         }
     },
     dob: {
-        type: String,
-        label: "Date of Birth"
-    },
-    locationId: {
-        type: String,
-        label: "Location"
-        //autoform: {
-        //    type: "select2",
-        //    afFieldInput: {
-        //        select2Options: {
-        //            // Must enter at least 1 character before we search
-        //            minimumInputLength: 1,
-        //            // Cannot select multiple items
-        //            multiple: false,
-        //            // Fetch the results with a Meteor method instead of ajax
-        //            ajax: {
-        //                transport: function (params, success, failure) {
-        //                    Meteor.call('school_listAddress', params.data.q, function (err, results) {
-        //                        if (err) {
-        //                            failure(err);
-        //                            return;
-        //                        }
-        //
-        //                        success(results);
-        //                    });
-        //                },
-        //                processResults: function (data) {
-        //                    var results = [];
-        //                    _.each(data.results, function (result) {
-        //                        results.push({
-        //                            id: result._id,
-        //                            text: result.name
-        //                        });
-        //                    });
-        //
-        //                    return {
-        //                        results: results
-        //                    };
-        //                }
-        //            }
-        //        }
-        //    }
-        //}
+        type: Date,
+        label: "Date of Birth",
+        defaultValue: function () {
+            return moment().format('YYYY-MM-DD');
+        }
     },
     telephone: {
         type: String,
@@ -79,6 +35,10 @@ Sample.Schema.Customer = new SimpleSchema({
         label: "Email",
         regEx: SimpleSchema.RegEx.Email,
         optional: true
+    },
+    locationId: {
+        type: String,
+        label: "Location"
     },
     photo: {
         type: String,
@@ -101,4 +61,4 @@ Sample.Schema.Customer = new SimpleSchema({
 Sample.Collection.Customer.attachSchema(Sample.Schema.Customer);
 
 // Attach soft remove
-Sample.Collection.Customer.attachBehaviour('softRemovable');
+//Sample.Collection.Customer.attachBehaviour('softRemovable');